requirements.txt
loguru==0.7.2
paho-mqtt==2.0.0
pyserial==3.5